+++ drivers/qspi/qspi.h
@@ -29,8 +29,8 @@ typedef struct
 	uint8_t		flags;
 	uint8_t		handle;
 	uint32_t	*private_data;
-	uint32_t	(*read)  (uint8_t handle, uint32_t address,uint8_t *data,uint16_t data_len);
-	uint32_t	(*write) (uint8_t handle, uint32_t address,uint8_t *data,uint16_t data_len);
+	uint32_t	(*read)  (uint8_t handle, uint32_t address,uint8_t *data,uint32_t data_len);
+	uint32_t	(*write) (uint8_t handle, uint32_t address,uint8_t *data,uint32_t data_len);
 	uint32_t	(*erase_blocks) (uint8_t handle, uint32_t start_block, uint32_t number_of_blocks);
 	uint32_t	(*erase_sectors) (uint8_t handle, uint32_t start_sector, uint32_t number_of_sectors);
 	uint32_t	(*erase_chip) (uint8_t handle);
@@ -42,10 +42,11 @@ typedef struct
 
 #ifdef STM32H7xx_HAL_QSPI_H
 #include "w25qxx/w25qxx.h"
+#include "w25qxx/w25qxx_defs.h"
 #endif
 
-extern	uint32_t qspi_read(uint8_t handle, uint32_t address,uint8_t *data,uint16_t data_len);
-extern	uint32_t qspi_write(uint8_t handle, uint32_t address,uint8_t *data,uint16_t data_len);
+extern	uint32_t qspi_read(uint8_t handle, uint32_t address,uint8_t *data,uint32_t data_len);
+extern	uint32_t qspi_write(uint8_t handle, uint32_t address,uint8_t *data,uint32_t data_len);
 extern	uint32_t qspi_erase_blocks(uint8_t handle, uint32_t start_block,uint32_t number_of_blocks);
 extern	uint32_t qspi_erase_sectors(uint8_t handle, uint32_t start_sector,uint32_t number_of_sectors);
 extern	uint32_t qspi_erase_chip(uint8_t handle);
